Added configResolver
The file contains two methods for resolving what configs to use on invocation of the program.

There are two types of config:
 + file
 + folder

The first type is defined in a .typicalrc file, which conforms to a specific structure, while the second is defined through a .typicalfolders directory, which contain a set of sub-directories that are taken as key-value pairs and added to the config.